+++ packages/components/src/components/action/button-pure/button-pure.tsx
@@ -1,5 +1,5 @@
-import type { ButtonAriaAttributes, PropTypes } from '../../../utils';
 import {
+  ALIGN_LABELS,
   AllowedTypes,
   attachComponentCss,
   BUTTON_ARIA_ATTRIBUTES,
@@ -10,6 +10,9 @@ import {
   improveButtonHandlingForCustomElement,
   isDisabledOrLoading,
   isSizeInherit,
+  TEXT_SIZES,
+  TEXT_WEIGHTS,
+  THEMES_EXTENDED_ELECTRIC_DARK,
   transitionListener,
   validateProps,
   warnIfParentIsPTextAndIconIsNone,
@@ -17,18 +20,18 @@ import {
 import type {
   AlignLabel,
   BreakpointCustomizable,
+  ButtonAriaAttributes,
   ButtonType,
   LinkButtonPureIconName,
+  PropTypes,
   SelectedAriaAttributes,
   TextSize,
   TextWeight,
   ThemeExtendedElectricDark,
 } from '../../../types';
-import { ALIGN_LABELS, TEXT_WEIGHTS, THEMES_EXTENDED_ELECTRIC_DARK } from '../../../types';
 import { Component, Element, h, Host, JSX, Listen, Prop } from '@stencil/core';
 import { getButtonAriaAttributes, warnIfIsLoadingAndIconIsNone } from './button-pure-utils';
 import { getComponentCss } from './button-pure-styles';
-import { TEXT_SIZES } from '../../basic/typography/text/text-utils';
 
 const propTypes: PropTypes<typeof ButtonPure> = {
   tabbable: AllowedTypes.boolean,