+++ packages/components/src/components/content/modal/modal-styles.ts
@@ -1,10 +1,10 @@
 import type { JssStyle } from 'jss';
-import type { BreakpointCustomizable, GetJssStyleFunction, BreakpointKey } from '../../../utils';
-import { mediaQueryMin } from '@porsche-design-system/utilities-v2';
+import type { GetJssStyleFunction } from '../../../utils';
+import type { BreakpointCustomizable, BreakpointKey } from '../../../types';
+import { gridSafeZone, mediaQueryMin } from '@porsche-design-system/utilities-v2';
 import { BREAKPOINTS, buildResponsiveStyles, getCss, mergeDeep, parseJSON, buildSlottedStyles } from '../../../utils';
 import {
   addImportantToEachRule,
-  contentWrapperVars,
   getBaseSlottedStyles,
   getFocusJssStyle,
   getInsetJssStyle,
@@ -32,10 +32,10 @@ export const getFullscreenJssStyles: GetJssStyleFunction = (fullscreen: boolean)
         margin: 0,
       }
     : {
-        minWidth: pxToRemWithUnit(272),
-        maxWidth: contentWrapperVars.maxWidth,
+        minWidth: pxToRemWithUnit(275.2), // 320px - 320px * 7% * 2
+        maxWidth: pxToRemWithUnit(1536), // 1920px - 1920px * 10% * 2
         minHeight: 'auto',
-        margin: `7vh ${contentWrapperVars.margin}`,
+        margin: `max(1rem, 7vh) ${gridSafeZone.base}`,
       };
 };
 
@@ -130,11 +130,11 @@ export const getComponentCss = (
           padding: pxToRemWithUnit(40),
         },
         [mediaQueryXl]: {
-          margin: isFullscreenForXlAndXxl ? 0 : `10vh ${contentWrapperVars.marginXl}`,
+          margin: isFullscreenForXlAndXxl ? 0 : `min(12rem, 10vh) ${gridSafeZone.xl}`,
         },
         [mediaQueryXxl]: {
           padding: pxToRemWithUnit(64),
-          margin: isFullscreenForXlAndXxl ? 0 : `10vh ${contentWrapperVars.marginXxl}`,
+          margin: isFullscreenForXlAndXxl ? 0 : `min(12rem, 10vh) ${gridSafeZone.xl}`,
         },
       },
       buildResponsiveStyles(fullscreen, getFullscreenJssStyles) as any