+++ packages/components/src/components/feedback/inline-notification/inline-notification.tsx
@@ -1,18 +1,17 @@
 import { JSX, Component, Prop, h, Element, Event, EventEmitter, Host } from '@stencil/core';
 import {
-  getPrefixedTagNames,
+  AllowedTypes,
   attachComponentCss,
   attachSlottedCss,
+  getPrefixedTagNames,
   hasHeading,
-  AllowedTypes,
+  THEMES,
   validateProps,
 } from '../../../utils';
-import type { PropTypes } from '../../../utils';
-import type { IconName, Theme } from '../../../types';
+import type { IconName, PropTypes, Theme } from '../../../types';
 import { getComponentCss, getSlottedCss } from './inline-notification-styles';
 import { INLINE_NOTIFICATION_STATES, getContentAriaAttributes, getIconName } from './inline-notification-utils';
 import type { InlineNotificationState } from './inline-notification-utils';
-import { THEMES } from '../../../types';
 
 const propTypes: PropTypes<typeof InlineNotification> = {
   heading: AllowedTypes.string,