+++ packages/components/src/components/form/select-wrapper/select-wrapper/select-wrapper.tsx
@@ -14,17 +14,16 @@ import {
   observeAttributes,
   setAriaAttributes,
   setAttribute,
+  THEMES,
   unobserveAttributes,
   validateProps,
 } from '../../../../utils';
-import type { PropTypes } from '../../../../utils';
-import type { BreakpointCustomizable, Theme } from '../../../../types';
+import type { BreakpointCustomizable, PropTypes, Theme } from '../../../../types';
 import type { DropdownDirection } from './select-wrapper-utils';
 import { DROPDOWN_DIRECTIONS, isCustomDropdown } from './select-wrapper-utils';
 import { getComponentCss, getSlottedCss } from './select-wrapper-styles';
 import { StateMessage } from '../../../common/state-message/state-message';
 import { Required } from '../../../common/required/required';
-import { THEMES } from '../../../../types';
 import { FORM_STATES } from '../../form-state';
 import type { FormState } from '../../form-state';