+++ packages/components/src/components/layout/flex/flex-item/flex-item-utils.ts
@@ -1,4 +1,4 @@
-import { BreakpointCustomizable } from '../../../../utils';
+import { BreakpointCustomizable } from '../../../../types';
 
 export const FLEX_ITEM_WIDTHS = [
   'auto',