+++ packages/components/src/utils/theme/isThemeDarkElectric.ts
@@ -1,0 +1,5 @@
+import { Theme } from '@porsche-design-system/utilities-v2';
+
+export const isThemeDarkElectric = (theme: Theme): boolean => {
+  return theme === 'dark-electric';
+};