+++ packages/components/src/utils/theme/theme-extended-electric.ts
@@ -1,0 +1,4 @@
+import { THEMES } from './theme';
+
+export const THEMES_EXTENDED_ELECTRIC = [...THEMES, 'light-electric'] as const;
+export type ThemeExtendedElectric = typeof THEMES_EXTENDED_ELECTRIC[number];