+++ packages/components-js/tests/e2e/helpers/puppeteer-helper.ts
@@ -274,6 +274,7 @@ export const initConsoleObserver = (page: Page): void => {
   });
 };
 export const getConsoleErrorsAmount = () => consoleMessages.filter((x) => x.type() === 'error').length;
+export const getConsoleWarningsAmount = () => consoleMessages.filter((x) => x.type() === 'warning').length;
 
 const thrownErrors: string[] = [];