+++ packages/components/CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### [Unreleased]
 
-### [2.15.0-beta.0] - 2022-08-02
+### [2.15.0-beta.0] - 2022-08-05
 
 #### Fixed